modified:   src/Pages/CreateNewProduct.tsx 	new file:   src/Pages/orders.tsx 	new file:   src/components/AdminRoute.tsx 	new file:   src/components/ProductFooter/ProductFooter.tsx 	renamed:    src/components/ReviewForm.tsx -> src/components/ProductFooter/ReviewForm.tsx 	renamed:    src/components/ReviewList.tsx -> src/components/ProductFooter/ReviewList.tsx 	renamed:    src/components/TabSelector.tsx -> src/components/ProductFooter/TabSelector.tsx 	modified:   src/components/Routes.tsx 	new file:   src/components/UserRoute.tsx 	deleted:    src/components/my-orders/orderTable.tsx 	deleted:    src/components/my-orders/orderTableBody.tsx 	renamed:    src/components/my-orders/orderRow.tsx -> src/components/orders/orderRow.tsx 	new file:   src/components/orders/orderTable.tsx 	new file:   src/components/orders/orderTableBody.tsx 	renamed:    src/components/my-orders/orderTableHeader.tsx -> src/components/orders/orderTableHeader.tsx 	deleted:    src/data/orders.ts 	deleted:    src/data/users.ts 	modified:   src/pages/Home.tsx
