+++ src/Pages/orders.tsx
@@ -1,0 +1,11 @@
+import OrderTable from "../components/orders/orderTable";
+
+function orders() {
+  return (
+    <>
+      <OrderTable />
+    </>
+  );
+}
+
+export default orders;
