+++ src/components/orders/orderTableBody.tsx
@@ -1,0 +1,15 @@
+// import OrderRow from "./orderRow.tsx";
+
+const OrderTableBody = () => {
+  return (
+    <tbody>
+      {/* {
+            orders.map((order) => (
+                // <OrderRow order={order} key={order.getId} />
+            ))
+        } */}
+    </tbody>
+  );
+};
+
+export default OrderTableBody;
